refactor: :recycle: Move instance definition to separate module
